backend/requirements.txt
eventlet==0.33.3
Flask==2.3.3
Flask-SocketIO==5.3.5
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.3
Flask-Cors==4.0.1
pytest==7.4.0
pymongo==4.5.0
bcrypt==4.0.1
PyJWT==2.8.0
google-cloud-vision==3.4.4
